Replaced aek3 with the adaptive sampling version 3b
